pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ventauri"
version = "0.1.0"
description = "Ventauri — multi-agent AI operating system for entrepreneurs"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Ventauri Contributors" }]
keywords = ["ai", "agents", "orchestration", "startup", "entrepreneur"]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "typer>=0.12.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "numpy>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    "jinja2>=3.1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
openai = ["openai>=1.55.0"]

[project.scripts]
ventauri = "cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = [
    "orchestrator*",
    "agents*",
    "memory*",
    "skills*",
    "workflows*",
    "api*",
    "cli*",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
